Refactor julekalender
